chore: remove dependency from api to unstructured
